Move TSDocTagDefinition.ts under src/configuration
